Migrated delete nom api call to modal and to submit form
